Implemented chain_rule

+++ minitorch/operators.py
@@ -13,22 +13,22 @@ import functools
 
 def mul(x: float, y: float) -> float:
     "$f(x, y) = x * y$"
-    return x * y
+    return float(x * y)
 
 
 def id(x: float) -> float:
     "$f(x) = x$"
-    return x
+    return float(x)
 
 
 def add(x: float, y: float) -> float:
     "$f(x, y) = x + y$"
-    return x + y
+    return float(x + y)
 
 
 def neg(x: float) -> float:
     "$f(x) = -x$"
-    return -x
+    return float(-x)
 
 
 def lt(x: float, y: float) -> float:
@@ -43,7 +43,7 @@ def eq(x: float, y: float) -> float:
 
 def max(x: float, y: float) -> float:
     "$f(x) =$ x if x is greater than y else y"
-    return x if x > y else y
+    return float(x if x > y else y)
 
 
 def is_close(x: float, y: float) -> float:
@@ -72,7 +72,7 @@ def relu(x: float) -> float:
 
     (See https://en.wikipedia.org/wiki/Rectifier_(neural_networks) .)
     """
-    return x if x > 0 else 0
+    return float(x) if x > 0 else 0.0
 
 
 EPS = 1e-6
@@ -136,7 +136,7 @@ def relu_back(x: float, d: float) -> float:
     Returns:
         float: The downstream gradient, equal to d if x > 0, else 0.
     """
-    return d if x > 0 else 0
+    return float(d) if x > 0 else 0.0
 
 
 # ## Task 0.3
